mca/threads: fix Pthreads mutex implementation.
