+++ opal/mca/threads/pthreads/threads_pthreads_mutex.c
@@ -31,16 +31,6 @@
 #include "opal/mca/threads/mutex.h"
 //#include "opal/mca/threads/pthreads/mutex_unix.h"
 
-
-OBJ_CLASS_INSTANCE(opal_mutex_t,
-                   opal_object_t,
-                   NULL,
-                   NULL);
-OBJ_CLASS_INSTANCE(opal_recursive_mutex_t,
-                   opal_object_t,
-                   NULL,
-                   NULL);
-
 /*
  * Wait and see if some upper layer wants to use threads, if support
  * exists.
@@ -63,3 +53,46 @@ struct opal_pthread_mutex_t {
 typedef struct opal_pthread_mutex_t opal_pthread_mutex_t;
 typedef struct opal_pthread_mutex_t opal_pthread_recursive_mutex_t;
 
+static void mca_threads_pthreads_mutex_constructor(opal_mutex_t *p_mutex) {
+    pthread_mutex_init(&p_mutex->m_lock_pthread, NULL);
+#if OPAL_ENABLE_DEBUG
+    p_mutex->m_lock_debug = 0;
+    p_mutex->m_lock_file = NULL;
+    p_mutex->m_lock_line = 0;
+#endif
+    opal_atomic_lock_init(&p_mutex->m_lock_atomic, 0);
+}
+
+static void mca_threads_pthreads_mutex_desctructor(opal_mutex_t *p_mutex) {
+    pthread_mutex_destroy(&p_mutex->m_lock_pthread);
+}
+
+static void mca_threads_pthreads_recursive_mutex_constructor
+        (opal_recursive_mutex_t *p_mutex) {
+    pthread_mutexattr_t mutex_attr;
+    pthread_mutexattr_init(&mutex_attr);
+    pthread_mutexattr_settype(&mutex_attr, PTHREAD_MUTEX_RECURSIVE);
+    pthread_mutex_init(&p_mutex->m_lock_pthread, &mutex_attr);
+    pthread_mutexattr_destroy(&mutex_attr);
+#if OPAL_ENABLE_DEBUG
+    p_mutex->m_lock_debug = 0;
+    p_mutex->m_lock_file = NULL;
+    p_mutex->m_lock_line = 0;
+#endif
+    opal_atomic_lock_init(&p_mutex->m_lock_atomic, 0);
+}
+
+static void mca_threads_pthreads_recursive_mutex_desctructor
+        (opal_recursive_mutex_t *p_mutex) {
+    pthread_mutex_destroy(&p_mutex->m_lock_pthread);
+}
+
+OBJ_CLASS_INSTANCE(opal_mutex_t,
+                   opal_object_t,
+                   mca_threads_pthreads_mutex_constructor,
+                   mca_threads_pthreads_mutex_desctructor);
+
+OBJ_CLASS_INSTANCE(opal_recursive_mutex_t,
+                   opal_object_t,
+                   mca_threads_pthreads_recursive_mutex_constructor,
+                   mca_threads_pthreads_recursive_mutex_desctructor);
